+++ api/orchestrator/orchestrator.proto
@@ -538,21 +538,23 @@ message Catalog {
 
   string description = 3;
 
-  repeated Category categories = 4;
+  repeated Category categories = 4
+      [ (tagger.tags) = "gorm:\"constraint:OnDelete:CASCADE\"" ];
 }
 
 message Category {
-
   string name = 1 [ (tagger.tags) = "gorm:\"primaryKey\"" ];
 
-  string description = 2;
+  // Reference to the catalog this category belongs to.
+  string catalog_id = 2 [ (tagger.tags) = "gorm:\"primaryKey\"" ];
 
-  repeated Control controls = 3
-      [ (tagger.tags) = "gorm:\"foreignKey:category_name,category_"
-                        "catalog_id;references:name,catalog_id\"" ];
+  string description = 3;
 
-  // Reference to the catalog this Category belongs to.
-  string catalog_id = 4 [ (tagger.tags) = "gorm:\"primaryKey\"" ];
+  repeated Control controls = 4 [
+    (tagger.tags) =
+        "gorm:\"foreignKey:category_name,category_"
+        "catalog_id;references:name,catalog_id;constraint:OnDelete:CASCADE\""
+  ];
 }
 
 // Control represents a certain Control that needs to be fulfilled. It could be
@@ -584,7 +586,8 @@ message Control {
   // the id field of the metric is populated - or a list of populated metric
   // meta-data, most likely returned by the database.
   repeated Metric metrics = 5
-      [ (tagger.tags) = "gorm:\"many2many:control_metrics\"" ];
+      [ (tagger.tags) =
+            "gorm:\"many2many:control_metrics;constraint:OnDelete:CASCADE\"" ];
 
   // Reference to the parent category this control belongs to.
