+++ internal/testutil/orchestratortest/orchestrator.go
@@ -8,6 +8,13 @@ import (
 	"google.golang.org/protobuf/types/known/structpb"
 )
 
+var (
+	MockCatalogID    = "Cat1234"
+	MockCategoryName = "My name"
+	MockControlID    = "Cont1234"
+	MockSubControlID = "Cont1234.1"
+)
+
 // NewCertificate creates a mock certificate
 func NewCertificate() *orchestrator.Certificate {
 	var mockCertificate = &orchestrator.Certificate{
@@ -30,30 +37,6 @@ func NewCertificate() *orchestrator.Certificate {
 	}
 
 	return mockCertificate
-}
-
-// NewCertificate creates a mock certificate
-func NewControl() *orchestrator.Control {
-	var mockControl = &orchestrator.Control{
-		ShortName:         "Cont1234",
-		CategoryName:      "My name",
-		CategoryCatalogId: "Cat1234",
-		Name:              "Mock Control",
-		Description:       "This is a mock control",
-		Metrics: []*assessment.Metric{{
-			Id:          "MockMetric",
-			Name:        "A Mock Metric",
-			Description: "This Metric is a mock metric",
-			Scale:       assessment.Metric_ORDINAL,
-			Range: &assessment.Range{
-				Range: &assessment.Range_AllowedValues{AllowedValues: &assessment.AllowedValues{
-					Values: []*structpb.Value{
-						structpb.NewBoolValue(false),
-						structpb.NewBoolValue(true),
-					}}}},
-		}},
-	}
-	return mockControl
 }
 
 // NewCatalog creates a mock catalog
@@ -83,10 +66,12 @@ func NewCatalog() *orchestrator.Catalog {
 							}}}},
 				}},
 				Controls: []*orchestrator.Control{{
-					ShortName:   "Cont1234.1",
-					Name:        "Mock Sub-Control",
-					Description: "This is a mock sub-control",
-					Metrics:     []*assessment.Metric{},
+					ShortName:         "Cont1234.1",
+					Name:              "Mock Sub-Control",
+					Description:       "This is a mock sub-control",
+					Metrics:           []*assessment.Metric{},
+					CategoryName:      "My name",
+					CategoryCatalogId: "Cat1234",
 				}},
 			}},
 		}}}
