+++ service/orchestrator/catalogs_test.go
@@ -2,8 +2,10 @@ package orchestrator
 
 import (
 	"context"
+	"fmt"
 	"testing"
 
+	"clouditor.io/clouditor/api"
 	"clouditor.io/clouditor/api/assessment"
 	"clouditor.io/clouditor/api/orchestrator"
 	"clouditor.io/clouditor/internal/testutil"
@@ -11,9 +13,285 @@ import (
 	"clouditor.io/clouditor/persistence"
 
 	"github.com/stretchr/testify/assert"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/types/known/structpb"
 )
+
+func Test_CreateCatalog(t *testing.T) {
+	// Mock catalogs
+	mockCatalog := orchestratortest.NewCatalog()
+	mockCatalogWithoutID := orchestratortest.NewCatalog()
+	mockCatalogWithoutID.Id = ""
+
+	type args struct {
+		in0 context.Context
+		req *orchestrator.CreateCatalogRequest
+	}
+	tests := []struct {
+		name         string
+		args         args
+		wantResponse *orchestrator.Catalog
+		wantErr      assert.ErrorAssertionFunc
+	}{
+		{
+			"missing request",
+			args{
+				context.Background(),
+				nil,
+			},
+			nil,
+			func(tt assert.TestingT, err error, i ...interface{}) bool {
+				assert.ErrorContains(t, err, api.ErrRequestIsNil.Error())
+				return assert.Equal(t, status.Code(err), codes.InvalidArgument)
+			},
+		},
+		{
+			"missing catalog",
+			args{
+				context.Background(),
+				&orchestrator.CreateCatalogRequest{},
+			},
+			nil,
+			func(tt assert.TestingT, err error, i ...interface{}) bool {
+				assert.ErrorContains(t, err, orchestrator.ErrCatalogIsNil.Error())
+				return assert.Equal(t, status.Code(err), codes.InvalidArgument)
+			},
+		},
+		{
+			"missing catalog id",
+			args{
+				context.Background(),
+				&orchestrator.CreateCatalogRequest{
+					Catalog: mockCatalogWithoutID,
+				},
+			},
+			nil,
+			func(tt assert.TestingT, err error, i ...interface{}) bool {
+				assert.ErrorContains(t, err, orchestrator.ErrCatalogIDIsMissing.Error())
+				return assert.Equal(t, status.Code(err), codes.InvalidArgument)
+			},
+		},
+		{
+			"valid catalog",
+			args{
+				context.Background(),
+				&orchestrator.CreateCatalogRequest{
+					Catalog: mockCatalog,
+				},
+			},
+			mockCatalog,
+			nil,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			s := NewService()
+			gotResponse, err := s.CreateCatalog(tt.args.in0, tt.args.req)
+
+			if err != nil && tt.wantErr != nil {
+				tt.wantErr(t, err)
+				return
+			} else {
+				assert.Nil(t, err)
+			}
+
+			// If no error is wanted, check response
+			if !proto.Equal(gotResponse, tt.wantResponse) {
+				t.Errorf("Service.CreateCatalog() = %v, want %v", gotResponse, tt.wantResponse)
+			}
+		})
+	}
+}
+
+func TestService_GetCatalog(t *testing.T) {
+	type fields struct {
+		storage persistence.Storage
+	}
+	type args struct {
+		req *orchestrator.GetCatalogRequest
+	}
+	tests := []struct {
+		name         string
+		fields       fields
+		args         args
+		wantResponse assert.ValueAssertionFunc
+		wantErr      assert.ErrorAssertionFunc
+	}{
+		{
+			name: "invalid request",
+			fields: fields{storage: testutil.NewInMemoryStorage(t, func(s persistence.Storage) {
+				// Create Catalog
+				assert.NoError(t, s.Create(orchestratortest.NewCatalog()))
+			})},
+			wantResponse: assert.Nil,
+			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
+				assert.Equal(t, codes.InvalidArgument, status.Code(err))
+				return assert.ErrorContains(t, err, api.ErrRequestIsNil.Error())
+			},
+		},
+		{
+			name: "catalog not found",
+			fields: fields{storage: testutil.NewInMemoryStorage(t, func(s persistence.Storage) {
+				// Create Catalog
+				assert.NoError(t, s.Create(orchestratortest.NewCatalog()))
+			})},
+			args:         args{req: &orchestrator.GetCatalogRequest{CatalogId: ""}},
+			wantResponse: assert.Nil,
+			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
+				assert.Equal(t, codes.NotFound, status.Code(err))
+				return assert.ErrorContains(t, err, "catalog ID is empty")
+			},
+		},
+		{
+			name: "valid",
+			fields: fields{storage: testutil.NewInMemoryStorage(t, func(s persistence.Storage) {
+				// Create Catalog
+				assert.NoError(t, s.Create(orchestratortest.NewCatalog()))
+			})},
+			args: args{req: &orchestrator.GetCatalogRequest{CatalogId: "Cat1234"}},
+			wantResponse: func(t assert.TestingT, i interface{}, i2 ...interface{}) bool {
+				res, ok := i.(*orchestrator.Catalog)
+				want := orchestratortest.NewCatalog()
+				assert.True(t, ok)
+				fmt.Println(res)
+				assert.Equal(t, 1, len(res.Categories))
+				return assert.Equal(t, want.Id, res.Id)
+			},
+			wantErr: assert.NoError,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			orchestratorService := Service{
+				storage: tt.fields.storage,
+			}
+			res, err := orchestratorService.GetCatalog(context.Background(), tt.args.req)
+
+			// Validate the error via the ErrorAssertionFunc function
+			tt.wantErr(t, err)
+
+			// Validate the response via the ValueAssertionFunc function
+			tt.wantResponse(t, res)
+		})
+	}
+}
+
+func Test_ListCatalogs(t *testing.T) {
+	var (
+		listCatalogsResponse *orchestrator.ListCatalogsResponse
+		err                  error
+	)
+
+	orchestratorService := NewService()
+	// 1st case: No services stored
+	listCatalogsResponse, err = orchestratorService.ListCatalogs(context.Background(), &orchestrator.ListCatalogsRequest{})
+	assert.NoError(t, err)
+	assert.NotNil(t, listCatalogsResponse.Catalogs)
+	assert.NotEmpty(t, listCatalogsResponse.Catalogs)
+
+	// 2nd case: One service stored
+	err = orchestratorService.storage.Create(orchestratortest.NewCatalog())
+	assert.NoError(t, err)
+
+	listCatalogsResponse, err = orchestratorService.ListCatalogs(context.Background(), &orchestrator.ListCatalogsRequest{})
+	assert.NoError(t, err)
+	assert.NotNil(t, listCatalogsResponse.Catalogs)
+	assert.NotEmpty(t, listCatalogsResponse.Catalogs)
+	// there is a default catalog plus the mock catalog
+	assert.Equal(t, len(listCatalogsResponse.Catalogs), 2)
+
+	// 3rd case: Invalid request
+	_, err = orchestratorService.ListCatalogs(context.Background(),
+		&orchestrator.ListCatalogsRequest{OrderBy: "not a field"})
+	assert.Equal(t, codes.InvalidArgument, status.Code(err))
+	assert.Contains(t, err.Error(), api.ErrInvalidColumnName.Error())
+}
+
+func Test_UpdateCatalog(t *testing.T) {
+	var (
+		catalog *orchestrator.Catalog
+		err     error
+	)
+	orchestratorService := NewService()
+
+	// 1st case: Certificate is nil
+	_, err = orchestratorService.UpdateCatalog(context.Background(), &orchestrator.UpdateCatalogRequest{})
+	assert.Equal(t, codes.InvalidArgument, status.Code(err))
+
+	// 2nd case: Certificate ID is nil
+	_, err = orchestratorService.UpdateCatalog(context.Background(), &orchestrator.UpdateCatalogRequest{
+		Catalog: catalog,
+	})
+	assert.Equal(t, codes.InvalidArgument, status.Code(err))
+
+	// 3rd case: Certificate not found since there are no certificates yet
+	_, err = orchestratorService.UpdateCatalog(context.Background(), &orchestrator.UpdateCatalogRequest{
+		Catalog: &orchestrator.Catalog{
+			Id: "Cat1234",
+		},
+		CatalogId: "Cat1234",
+	})
+	assert.Equal(t, codes.NotFound, status.Code(err))
+
+	// 4th case: Certificate updated successfully
+	mockCatalog := orchestratortest.NewCatalog()
+	err = orchestratorService.storage.Create(mockCatalog)
+	assert.NoError(t, err)
+	if err != nil {
+		return
+	}
+
+	// update the certificate's description and send the update request
+	mockCatalog.Description = "new description"
+	catalog, err = orchestratorService.UpdateCatalog(context.Background(), &orchestrator.UpdateCatalogRequest{
+		CatalogId: "Cat1234",
+		Catalog:   mockCatalog,
+	})
+	assert.NoError(t, err)
+	assert.NotNil(t, catalog)
+	assert.Equal(t, "new description", catalog.Description)
+}
+
+func Test_RemoveCatalog(t *testing.T) {
+	var (
+		err                  error
+		listCatalogsResponse *orchestrator.ListCatalogsResponse
+	)
+	orchestratorService := NewService()
+
+	// 1st case: Empty certificate ID error
+	_, err = orchestratorService.RemoveCatalog(context.Background(), &orchestrator.RemoveCatalogRequest{CatalogId: ""})
+	assert.Error(t, err)
+	assert.Equal(t, status.Code(err), codes.InvalidArgument)
+
+	// 2nd case: ErrRecordNotFound
+	_, err = orchestratorService.RemoveCatalog(context.Background(), &orchestrator.RemoveCatalogRequest{CatalogId: "0000"})
+	assert.Error(t, err)
+	assert.Equal(t, status.Code(err), codes.NotFound)
+
+	// 3rd case: Record removed successfully
+	mockCatalog := orchestratortest.NewCatalog()
+	err = orchestratorService.storage.Create(mockCatalog)
+	assert.NoError(t, err)
+
+	// There is a record for catalogs in the DB (default one)
+	listCatalogsResponse, err = orchestratorService.ListCatalogs(context.Background(), &orchestrator.ListCatalogsRequest{})
+	assert.NoError(t, err)
+	assert.NotNil(t, listCatalogsResponse.Catalogs)
+	assert.Equal(t, 2, len(listCatalogsResponse.Catalogs))
+
+	// Remove record
+	_, err = orchestratorService.RemoveCatalog(context.Background(), &orchestrator.RemoveCatalogRequest{CatalogId: mockCatalog.Id})
+	assert.NoError(t, err)
+
+	// There is a record for cloud services in the DB (default one)
+	listCatalogsResponse, err = orchestratorService.ListCatalogs(context.Background(), &orchestrator.ListCatalogsRequest{})
+	assert.NoError(t, err)
+	assert.NotNil(t, listCatalogsResponse.Catalogs)
+	assert.Equal(t, 1, len(listCatalogsResponse.Catalogs))
+}
 
 func TestService_GetCategory(t *testing.T) {
 	type fields struct {
@@ -117,7 +395,7 @@ func TestService_GetControl(t *testing.T) {
 			fields: fields{
 				storage: testutil.NewInMemoryStorage(t, func(s persistence.Storage) {
 					// Create Catalog
-					assert.NoError(t, s.Create(orchestratortest.NewControl()))
+					assert.NoError(t, s.Create(orchestratortest.NewCatalog()))
 				})},
 			args: args{req: &orchestrator.GetControlRequest{CatalogId: "Cat1234", CategoryName: "My name", ControlShortName: "Cont1234"}},
 			wantRes: &orchestrator.Control{
@@ -137,6 +415,17 @@ func TestService_GetControl(t *testing.T) {
 								structpb.NewBoolValue(false),
 								structpb.NewBoolValue(true),
 							}}}},
+				}},
+				Controls: []*orchestrator.Control{{
+					ShortName:                      "Cont1234.1",
+					Name:                           "Mock Sub-Control",
+					Description:                    "This is a mock sub-control",
+					Metrics:                        []*assessment.Metric{},
+					CategoryName:                   "My name",
+					CategoryCatalogId:              "Cat1234",
+					ParentControlShortName:         &orchestratortest.MockControlID,
+					ParentControlCategoryCatalogId: &orchestratortest.MockCatalogID,
+					ParentControlCategoryName:      &orchestratortest.MockCategoryName,
 				}},
 			},
 			wantErr: false,
